# docs/swagger.yaml
basePath: /
definitions:
...
  api_handler.ErrorResponse:
    properties:
      errors:
        items:
          $ref: '#/definitions/api_handler.FieldError'
        type: array
      message:
        example: Invalid request
        type: string
      status:
        example: 400
        type: integer
    type: object
...
  api_handler.FieldError:
    properties:
      field:
        example: email
        type: string
      message:
        example: Email is required
        type: string
    type: object
...
  api_handler.MessageResponse:
    properties:
      message:
        example: Operation completed successfully
        type: string
      status:
        example: 200
        type: integer
    type: object
  api_handler.PaginatedResponse:
    properties:
      data:
        items:
          type: object
        type: array
      message:
        example: Data retrieved successfully
        type: string
      pagination:
        $ref: '#/definitions/api_handler.PaginationInfo'
      status:
        example: 200
        type: integer
    type: object
  api_handler.PaginationInfo:
    properties:
      has_next:
        example: true
        type: boolean
      has_prev:
        example: false
        type: boolean
      page:
        example: 1
        type: integer
      page_size:
        example: 10
        type: integer
      total_count:
        example: 100
        type: integer
      total_pages:
        example: 10
        type: integer
    type: object
...
  api_handler.SuccessResponse:
    properties:
      data:
        type: object
      message:
        example: Operation completed successfully
        type: string
      status:
        example: 200
        type: integer
    type: object
...
  github_com_edsonmubezi_myapp_internal_organization.Organization:
    properties:
      address:
        minLength: 6
        type: string
      contact_person:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: integer
      name:
        maxLength: 100
        minLength: 3
        type: string
      phone_number:
        type: string
      updated_at:
        type: string
    required:
    - address
    - contact_person
    - name
    - phone_number
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@ekklesiahr.com
    name: EkklesiaHR API Support
    url: https://github.com/edsonmubezi/ekklesiahr
  description: Production-ready HR Management System API with comprehensive employee,
    payroll, leave management, and organization administration. Supports multi-tenant
    architecture with role-based access control.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: EkklesiaHR API
  version: 1.0.1
paths:
...
  /api/all-organizations:
    get:
      description: Retrieves paginated list of organizations with support for searching
        and sorting by various fields
      parameters:
      - default: 1
        description: Page number
        example: 1
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        example: 10
        in: query
        name: page_size
        type: integer
      - default: id
        description: Sort field
        enum:
        - id
        - name
        - created_at
        in: query
        name: sort_by
        type: string
      - default: asc
        description: Sort order
        enum:
        - asc
        - desc
        in: query
        name: sort_order
        type: string
      - description: Search by organization name
        example: '"Acme"'
        in: query
        name: q
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Organizations retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/api_handler.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_edsonmubezi_myapp_internal_organization.Organization'
                  type: array
              type: object
        "500":
          description: Failed to fetch organizations
          schema:
            $ref: '#/definitions/api_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List Organizations (Paginated)
      tags:
      - Organizations
...
  /api/organizations:
    get:
      description: Retrieves all organizations without pagination. For paginated results,
        use /api/all-organizations endpoint instead.
      produces:
      - application/json
      responses:
        "200":
          description: All organizations retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/api_handler.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_edsonmubezi_myapp_internal_organization.Organization'
                  type: array
              type: object
        "500":
          description: Failed to fetch organizations
          schema:
            $ref: '#/definitions/api_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List All Organizations (Unpaginated)
      tags:
      - Organizations
    post:
      consumes:
      - application/json
      description: Registers a new organization in the multi-tenant system with contact
        details and configuration settings
      parameters:
      - description: Organization details
        in: body
        name: organization
        required: true
        schema:
          $ref: '#/definitions/github_com_edsonmubezi_myapp_internal_organization.Organization'
      produces:
      - application/json
      responses:
        "201":
          description: Organization created successfully
          schema:
            allOf:
            - $ref: '#/definitions/api_handler.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_edsonmubezi_myapp_internal_organization.Organization'
              type: object
        "400":
          description: Invalid input or validation error
          schema:
            $ref: '#/definitions/api_handler.ErrorResponse'
        "500":
          description: Failed to create organization
          schema:
            $ref: '#/definitions/api_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create Organization
      tags:
      - Organizations
  /api/organizations/{id}:
    delete:
      consumes:
      - application/json
      description: Soft-deletes an organization by ID. Organization data is retained
        in database but marked as deleted and excluded from queries.
      parameters:
      - description: Organization ID
        example: 1
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Organization deleted successfully
          schema:
            $ref: '#/definitions/api_handler.MessageResponse'
        "400":
          description: Invalid organization ID
          schema:
            $ref: '#/definitions/api_handler.ErrorResponse'
        "500":
          description: Failed to delete organization
          schema:
            $ref: '#/definitions/api_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete Organization
      tags:
      - Organizations
    get:
      description: Retrieves complete organization details including name, contact
        information, and configuration for a specific organization
      parameters:
      - description: Organization ID
        example: 1
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Organization details retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/api_handler.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_edsonmubezi_myapp_internal_organization.Organization'
              type: object
        "400":
          description: Invalid organization ID format
          schema:
            $ref: '#/definitions/api_handler.ErrorResponse'
        "404":
          description: Organization not found
          schema:
            $ref: '#/definitions/api_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Organization by ID
      tags:
      - Organizations
    put:
      consumes:
      - application/json
      description: Updates an existing organization's details including name, address,
        contact person, and configuration settings
      parameters:
      - description: Organization ID
        example: 1
        in: path
        name: id
        required: true
        type: integer
      - description: Updated organization details
        in: body
        name: organization
        required: true
        schema:
          $ref: '#/definitions/github_com_edsonmubezi_myapp_internal_organization.Organization'
      produces:
      - application/json
      responses:
        "200":
          description: Organization updated successfully
          schema:
            allOf:
            - $ref: '#/definitions/api_handler.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_edsonmubezi_myapp_internal_organization.Organization'
              type: object
        "400":
          description: Invalid organization ID or request body
          schema:
            $ref: '#/definitions/api_handler.ErrorResponse'
        "404":
          description: Organization not found
          schema:
            $ref: '#/definitions/api_handler.ErrorResponse'
        "500":
          description: Failed to update organization
          schema:
            $ref: '#/definitions/api_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update Organization
      tags:
      - Organizations
  /api/organizations/encrypted:
    get:
      description: Retrieves all organizations with ID encryption enabled. Organization
        IDs are encrypted in the response for additional security.
      produces:
      - application/json
      responses:
        "200":
          description: Organizations retrieved successfully with encrypted IDs
          schema:
            allOf:
            - $ref: '#/definitions/api_handler.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_edsonmubezi_myapp_internal_organization.Organization'
                  type: array
              type: object
        "500":
          description: Failed to fetch organizations
          schema:
            $ref: '#/definitions/api_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List Organizations with Encrypted IDs
      tags:
      - Organizations
...
schemes:
- https
- http
securityDefinitions:
  BearerAuth:
    description: 'Provide your JWT token as: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"